backend/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.4.0
SQLAlchemy==1.3.3
Flask-Cors==3.0.8
Flask-Migrate==2.6.0
Flask-Script==2.0.6
python-jose==3.2.0
six==1.12.0
requests==2.25.1
authlib==0.15.3